Add separate class files for Dataset class.
